+++ src/imagetransforms/ImageTransform.php
@@ -120,6 +120,9 @@ abstract class ImageTransform implements ImageTransformInterface
             } else {
                 try {
                     $url = UrlHelper::siteUrl($url, null, $protocol);
+                    if (UrlHelper::isProtocolRelativeUrl($url)) {
+                        $url = UrlHelper::urlWithScheme($url, $protocol);
+                    }
                 } catch (Exception $e) {
                 }
             }